@@ -9,4 +9,4 @@ if danger.git.createdFiles.count + danger.git.modifiedFiles.count - danger.git.d
     warn("Big PR, try to keep changes smaller if you can")
 }
 
-SwiftLint.lint(.all(directory: "Sources/CrowdinSDK"), configFile: "Sources/CrowdinSDK/.swiftlint.yml")
+SwiftLint.lint(.all(directory: "Sources/CrowdinSDK"), configFile: ".swiftlint.yml")
