@@ -10,7 +10,7 @@ import BaseAPI
 
 class StorageAPI: CrowdinAPI {
     fileprivate enum StorageRequestHeaderFields: String {
-        case CrowdinAPIFileName = "Crowdin-API-FileName"
+        case crowdinAPIFileName = "Crowdin-API-FileName"
     }
 
     override var apiPath: String {
@@ -21,7 +21,7 @@ class StorageAPI: CrowdinAPI {
         let apiFileName = fileName ?? String(Date().timeIntervalSince1970)
         let apiFileNameWithExtension = apiFileName.hasSuffix(".png") ? apiFileName : apiFileName + ".png"
         let headers = [RequestHeaderFields.contentType.rawValue: "image/png",
-                       StorageRequestHeaderFields.CrowdinAPIFileName.rawValue: apiFileNameWithExtension]
+                       StorageRequestHeaderFields.crowdinAPIFileName.rawValue: apiFileNameWithExtension]
         self.cw_post(url: fullPath, headers: headers, body: data, completion: completion)
     }
 }
