@@ -14,14 +14,11 @@ import Foundation
 
 protocol RealtimeUpdateFeatureProtocol {
     static var shared: RealtimeUpdateFeatureProtocol? { get set }
-
     var success: (() -> Void)? { get set }
     var error: ((Error) -> Void)? { set get }
     var disconnect: (() -> Void)? { get set }
     var enabled: Bool { get set }
-
-	init(hash: String, sourceLanguage: String, organizationName: String?, loginFeature: AnyLoginFeature?)
-
+	init(hash: String, sourceLanguage: String, organizationName: String?, minimumManifestUpdateInterval: TimeInterval, loginFeature: AnyLoginFeature?)
     func start()
     func stop()
     func subscribe(control: Refreshable)
@@ -39,10 +36,10 @@ class RealtimeUpdateFeature: RealtimeUpdateFeatureProtocol {
         let localizations = Localization.current.provider.remoteStorage.localizations
         return CrowdinSDK.currentLocalization ?? Bundle.main.preferredLanguage(with: localizations)
     }
-    var hashString: String
+    let hashString: String
     let sourceLanguage: String
     let organizationName: String?
-
+    let minimumManifestUpdateInterval: TimeInterval
 	var distributionResponse: DistributionsResponse? = nil
 
     var active: Bool { return socketManger?.active ?? false }
@@ -59,13 +56,13 @@ class RealtimeUpdateFeature: RealtimeUpdateFeatureProtocol {
     private var socketManger: CrowdinSocketManagerProtocol?
     private var mappingManager: CrowdinMappingManagerProtocol
     private let loginFeature: AnyLoginFeature?
-
-    required init(hash: String, sourceLanguage: String, organizationName: String?, loginFeature: AnyLoginFeature?) {
+    required init(hash: String, sourceLanguage: String, organizationName: String?, minimumManifestUpdateInterval: TimeInterval, loginFeature: AnyLoginFeature?) {
         self.hashString = hash
         self.sourceLanguage = sourceLanguage
 		self.organizationName = organizationName
         self.loginFeature = loginFeature
-        self.mappingManager = CrowdinMappingManager(hash: hash, sourceLanguage: sourceLanguage, organizationName: organizationName)
+        self.minimumManifestUpdateInterval = minimumManifestUpdateInterval
+        self.mappingManager = CrowdinMappingManager.shared(hash: hash, sourceLanguage: sourceLanguage, organizationName: organizationName, minimumManifestUpdateInterval: minimumManifestUpdateInterval)
     }
 
     func downloadDistribution(with successHandler: (() -> Void)? = nil, errorHandler: ((Error) -> Void)? = nil) {
@@ -115,16 +112,20 @@ class RealtimeUpdateFeature: RealtimeUpdateFeatureProtocol {
         }
     }
 
+    // swiftlint:disable identifier_name
     func _start() {
-		guard let projectId = distributionResponse?.data.project.id, let projectWsHash = distributionResponse?.data.project.wsHash, let userId = distributionResponse?.data.user.id, let wsUrl = distributionResponse?.data.wsUrl else {
+        guard let projectId = distributionResponse?.data.project.id,
+              let projectWsHash = distributionResponse?.data.project.wsHash,
+              let userId = distributionResponse?.data.user.id,
+              let wsUrl = distributionResponse?.data.wsUrl else {
             self.downloadDistribution(with: {
                 self._start()
             }, errorHandler: error)
-			return
+            return
 		}
         setupRealtimeUpdatesLocalizationProvider(with: projectId) { [weak self] in
             guard let self = self else { return }
-            self.setupSocketManager(with: projectId, projectWsHash: projectWsHash, userId: userId, wsUrl: wsUrl)
+            self.setupSocketManager(with: projectId, projectWsHash: projectWsHash, userId: userId, wsUrl: wsUrl, minimumManifestUpdateInterval: minimumManifestUpdateInterval)
         }
     }
 
@@ -137,11 +138,23 @@ class RealtimeUpdateFeature: RealtimeUpdateFeatureProtocol {
     }
 
     var oldProvider: LocalizationProviderProtocol? = nil
-
+    
     func setupRealtimeUpdatesLocalizationProvider(with projectId: String, completion: @escaping () -> Void) {
         oldProvider = Localization.current.provider
-        Localization.current.provider = LocalizationProvider(localization: self.localization, localStorage: RULocalLocalizationStorage(localization: self.localization), remoteStorage: RURemoteLocalizationStorage(localization: self.localization, sourceLanguage: sourceLanguage, hash: self.hashString, projectId: projectId, organizationName: self.organizationName))
-
+        let localStorage = RULocalLocalizationStorage(localization: self.localization)
+        let remoteStorage = RURemoteLocalizationStorage(
+            localization: self.localization,
+            sourceLanguage: sourceLanguage,
+            hash: self.hashString,
+            projectId: projectId,
+            organizationName: self.organizationName,
+            minimumManifestUpdateInterval: self.minimumManifestUpdateInterval
+        )
+        Localization.current.provider = LocalizationProvider(
+            localization: self.localization,
+            localStorage: localStorage,
+            remoteStorage: remoteStorage
+        )
         Localization.current.provider.refreshLocalization { [weak self] error in
             guard let self = self else { return }
             if let error = error {
@@ -166,19 +179,32 @@ class RealtimeUpdateFeature: RealtimeUpdateFeatureProtocol {
             self.refreshAllControls()
         }
     }
-
-    func setupSocketManager(with projectId: String, projectWsHash: String, userId: String, wsUrl: String) {
+    func setupSocketManager(with projectId: String, projectWsHash: String, userId: String, wsUrl: String, minimumManifestUpdateInterval: TimeInterval) {
         // Download manifest if it is not initialized.
-        let manifestManager = ManifestManager.manifest(for: hashString, sourceLanguage: sourceLanguage, organizationName: organizationName)
-        guard manifestManager.downloaded else {
+        let manifestManager = ManifestManager.manifest(
+            for: hashString,
+            sourceLanguage: sourceLanguage,
+            organizationName: organizationName,
+            minimumManifestUpdateInterval: minimumManifestUpdateInterval
+        )
+        guard manifestManager.available else {
             manifestManager.download { [weak self] in
                 guard let self = self else { return }
-                self.setupSocketManager(with: projectId, projectWsHash: projectWsHash, userId: userId, wsUrl: wsUrl)
+                self.setupSocketManager(with: projectId, projectWsHash: projectWsHash, userId: userId, wsUrl: wsUrl, minimumManifestUpdateInterval: minimumManifestUpdateInterval)
             }
             return
         }
 
-        self.socketManger = CrowdinSocketManager(hashString: hashString, projectId: projectId, projectWsHash: projectWsHash, userId: userId, wsUrl: wsUrl, languageResolver: manifestManager)
+        self.socketManger = CrowdinSocketManager(
+            hashString: hashString,
+            projectId: projectId,
+            projectWsHash: projectWsHash,
+            userId: userId,
+            wsUrl: wsUrl,
+            languageResolver: manifestManager,
+            organizationName: organizationName,
+            auth: loginFeature
+        )
         self.socketManger?.didChangeString = { id, newValue in
             self.didChangeString(with: id, to: newValue)
         }
