@@ -11,17 +11,19 @@ import Foundation
     var hashString: String
     var sourceLanguage: String
     var organizationName: String?
-
-    public init(hashString: String, sourceLanguage: String, organizationName: String? = nil) {
+    var minimumManifestUpdateInterval: TimeInterval
+    public init(hashString: String, sourceLanguage: String, organizationName: String? = nil, minimumManifestUpdateInterval: TimeInterval = Constants.defaultMinimumManifestUpdateInterval) {
         self.hashString = hashString
         self.sourceLanguage = sourceLanguage
         self.organizationName = organizationName
+        self.minimumManifestUpdateInterval = minimumManifestUpdateInterval
     }
 
     @available(*, deprecated, renamed: "init(hashString:sourceLanguage:)")
     public init(hashString: String, localizations: [String], sourceLanguage: String) {
         self.hashString = hashString
         self.sourceLanguage = sourceLanguage
+        self.minimumManifestUpdateInterval = Constants.defaultMinimumManifestUpdateInterval
     }
 
     public override init() {
@@ -33,5 +35,10 @@ import Foundation
             fatalError("Please add CrowdinPluralsFileNames key to your Info.plist file")
         }
         self.sourceLanguage = crowdinSourceLanguage
+        self.minimumManifestUpdateInterval = Constants.defaultMinimumManifestUpdateInterval
+    }
+    public enum Constants {
+        // New default minimum interval for manifest updates
+        public static let defaultMinimumManifestUpdateInterval: TimeInterval = 15 * 60 // 15 minutes
     }
 }
