@@ -17,6 +17,9 @@ class CrowdinLocalizationDownloader: CrowdinDownloaderProtocol {
     fileprivate var errors: [Error]? = nil
     fileprivate var contentDeliveryAPI: CrowdinContentDeliveryAPI!
     fileprivate let manifestManager: ManifestManager
+    
+    // Add a lock to protect shared resources
+    fileprivate let lock = NSLock()
 
     init(manifestManager: ManifestManager) {
         self.manifestManager = manifestManager
@@ -27,29 +30,48 @@ class CrowdinLocalizationDownloader: CrowdinDownloaderProtocol {
         self.getFiles(for: localization) { [weak self] (files, timestamp, error) in
             guard let self = self else { return }
             if let files = files {
-                let strings = files.filter({ $0.isStrings })
-                let plurals = files.filter({ $0.isStringsDict })
-                let xliffs = files.filter({ $0.isXliff })
-                let xcstrings = files.filter({ $0.isXcstrings })
-                self.download(strings: strings, plurals: plurals, xliffs:xliffs, xcstrings: xcstrings, with: hash, timestamp: timestamp, for: localization)
+                let xcstringsFiles = files.filter({ $0.isXcstrings })
+                let notXcstringsFiles = files.filter({ !$0.isXcstrings })
+                let notXcstringsFilesToDownload = notXcstringsFiles.filter { self.manifestManager.hasFileChanged(filePath: $0, localization: localization) }
+                let xcStringsFilesToDownlaod = xcstringsFiles.filter({ self.manifestManager.hasFileChanged(filePath: $0, localization: self.manifestManager.xcstringsLanguage) })
+                let filesToDownload = xcStringsFilesToDownlaod + notXcstringsFilesToDownload
+                if !filesToDownload.isEmpty {
+                    self.download(strings: filesToDownload.filter({ $0.isStrings }),
+                                  plurals: filesToDownload.filter({ $0.isStringsDict }),
+                                  xliffs: filesToDownload.filter({ $0.isXliff }),
+                                  xcstrings: filesToDownload.filter({ $0.isXcstrings }),
+                                  with: hash, timestamp: timestamp, for: localization)
+                } else {
+                    self.completion?(nil, nil, nil)
+                }
             } else if let error = error {
+                self.lock.lock()
                 self.errors = [error]
                 self.completion?(nil, nil, self.errors)
+                self.lock.unlock()
             }
         }
     }
 
     func download(strings: [String], plurals: [String], xliffs: [String], xcstrings: [String], with hash: String, timestamp: TimeInterval?, for localization: String) {
+        let timestamp = timestamp ?? Date().timeIntervalSince1970
         self.operationQueue.cancelAllOperations()
 
         self.contentDeliveryAPI = CrowdinContentDeliveryAPI(hash: hash, session: URLSession.shared)
+        
+        // Initialize shared resources with lock protection
+        lock.lock()
         self.strings = nil
         self.plurals = nil
         self.errors = nil
+        lock.unlock()
 
         let completionBlock = BlockOperation { [weak self] in
             guard let self = self else { return }
+            // Access shared resources with lock protection
+            self.lock.lock()
             self.completion?(self.strings, self.plurals, self.errors)
+            self.lock.unlock()
         }
 
         strings.forEach { filePath in
@@ -58,6 +80,9 @@ class CrowdinLocalizationDownloader: CrowdinDownloaderProtocol {
                 guard let self = self else { return }
                 self.add(strings: strings)
                 self.add(error: error)
+                if error == nil {
+                    self.updateTimestamp(for: localization, filePath: filePath, timestamp: timestamp)
+                }
             }
             completionBlock.addDependency(download)
             operationQueue.addOperation(download)
@@ -69,6 +94,9 @@ class CrowdinLocalizationDownloader: CrowdinDownloaderProtocol {
                 guard let self = self else { return }
                 self.add(plurals: plurals)
                 self.add(error: error)
+                if error == nil {
+                    self.updateTimestamp(for: localization, filePath: filePath, timestamp: timestamp)
+                }
             }
             completionBlock.addDependency(download)
             operationQueue.addOperation(download)
@@ -81,6 +109,9 @@ class CrowdinLocalizationDownloader: CrowdinDownloaderProtocol {
                 self.add(strings: strings)
                 self.add(plurals: plurals)
                 self.add(error: error)
+                if error == nil {
+                    self.updateTimestamp(for: localization, filePath: filePath, timestamp: timestamp)
+                }
             }
             completionBlock.addDependency(download)
             operationQueue.addOperation(download)
@@ -97,6 +128,9 @@ class CrowdinLocalizationDownloader: CrowdinDownloaderProtocol {
                 self.add(strings: strings)
                 self.add(plurals: plurals)
                 self.add(error: error)
+                if error == nil {
+                    self.updateTimestamp(for: self.manifestManager.xcstringsLanguage, filePath: filePath, timestamp: timestamp)
+                }
             }
             completionBlock.addDependency(download)
             operationQueue.addOperation(download)
@@ -122,28 +156,38 @@ class CrowdinLocalizationDownloader: CrowdinDownloaderProtocol {
 
     func add(error: Error?) {
         guard let error = error else { return }
+        lock.lock()
         if self.errors != nil {
             self.errors?.append(error)
         } else {
             self.errors = [error]
         }
+        lock.unlock()
     }
 
     func add(strings: [String: String]?) {
         guard let strings = strings else { return }
+        lock.lock()
         if self.strings != nil {
             self.strings?.merge(with: strings)
         } else {
             self.strings = strings
         }
+        lock.unlock()
     }
 
     func add(plurals: [AnyHashable: Any]?) {
         guard let plurals = plurals else { return }
+        lock.lock()
         if self.plurals != nil {
             self.plurals?.merge(with: plurals)
         } else {
             self.plurals = plurals
         }
+        lock.unlock()
+    }
+    func updateTimestamp(for localization: String, filePath: String, timestamp: TimeInterval) {
+        manifestManager.fileTimestampStorage.updateTimestamp(for: localization, filePath: filePath, timestamp: timestamp)
+        manifestManager.fileTimestampStorage.saveTimestamps()
     }
 }
